refactor background image compomnent
