--- src/client/core/components/BackgroundImage/background-image.scss
@@ -1,10 +1,0 @@
-@import '~styles/config/_colors.scss';
-
-.background-image {
-  &__container {
-    &__image {
-      height: auto;
-      width: 100%;
-    }
-  }
-}
