+++ src/client/core/components/ImageComponent/image.scss
@@ -1,0 +1,8 @@
+@import '~styles/config/_colors.scss';
+
+.image__container {
+  &__image {
+    height: auto;
+    width: 100%;
+  }
+}
