+++ src/client/core/modules/Hero/hero.scss
@@ -21,6 +21,7 @@
       position: absolute;
       width: 100%;
       box-sizing: border-box;
+      user-select: none;
       &__headline {
         font-size: 75px;
         line-height: 75px;
